back_requirements.txt
discord.py==2.0.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
urllib3==1.26.8